Foil: cut the library over to the new checker
Rzk.TypeCheck is now a façade over the layers; Free/Scoped.hs,
Free/Scoped/TH.hs, Language/Rzk/Free/Syntax.hs and the 5,500-line
Rzk/TypeCheck.hs are gone. Language.Rzk.Foil.Names keeps what survives of
the old syntax module -- surface identifiers, binders and modalities --
which is what the checker shows the user and is independent of how a term
represents its variables.

The LSP's incremental cache changes shape. It used to hold each module's
elaborated declarations and replay them into a fresh context; a cached
term now names the definitions it uses by their foil name, which only
means anything in the scope that produced it, so the cache holds the
context of each checked prefix and a re-check resumes from it. What the
LSP displays -- completion, symbols, hover -- reads a rendered DeclView
instead, and never touches a scope-indexed term.

The reference index elaborated the annotation of a pair binder through the
old core to split it (giving hover 'q : B p' rather than 'q : B x'). The
annotation is an open term, so it now goes through withOpenTerm, which
gives every identifier in it a name and maps them back when printing.
