+++ rzk/src/Rzk/TypeCheck/BinderTypes.hs
@@ -27,7 +27,7 @@ import           Control.Monad.Foil       (Distinct)
 import           Control.Monad.Free.Foil  (AST (Node, Var), ScopedAST (..))
 
 import           Language.Rzk.Foil.Syntax
-import           Language.Rzk.Free.Syntax (Binder (..), TModality (..),
+import           Language.Rzk.Foil.Names (Binder (..), TModality (..),
                                            TypeInfo (..), VarIdent)
 import           Rzk.TypeCheck.Context
 import           Rzk.TypeCheck.Decl
