+++ rzk/src/Rzk/TypeCheck/Context.hs
@@ -51,7 +51,7 @@ import qualified Data.Set                    as Set
 import           Unsafe.Coerce               (unsafeCoerce)
 
 import           Language.Rzk.Foil.Syntax
-import           Language.Rzk.Free.Syntax    (Binder (..), TModality (..),
+import           Language.Rzk.Foil.Names    (Binder (..), TModality (..),
                                               VarIdent, binderName)
 import qualified Language.Rzk.Syntax         as Rzk
