+++ rzk/src/Rzk/TypeCheck/Judgements.hs
@@ -29,7 +29,7 @@ import qualified Control.Monad.Foil       as Foil
 import           Control.Monad.Free.Foil  (AST (Var), ScopedAST (..))
 
 import           Language.Rzk.Foil.Syntax
-import           Language.Rzk.Free.Syntax (Binder (..), TModality (..),
+import           Language.Rzk.Foil.Names (Binder (..), TModality (..),
                                            TypeInfo (..), VarIdent,
                                            binderDisplayName, binderIsCompound,
                                            binderLeaves, binderName,
