+++ rzk/src/Rzk/TypeCheck/Render.hs
@@ -25,7 +25,7 @@ import qualified Control.Monad.Foil       as Foil
 import           Control.Monad.Free.Foil  (AST (Var))
 
 import           Language.Rzk.Foil.Syntax
-import           Language.Rzk.Free.Syntax (Binder (..), TModality (..),
+import           Language.Rzk.Foil.Names (Binder (..), TModality (..),
                                            binderName)
 import           Rzk.Render.Geometry
 import           Rzk.TypeCheck.Context
